requirements/base.txt
# ***** Django libraries ******
Django==3.1.5
django-auto-prefetching==0.1.8
django-cacheops==5.1
django-celery-beat==2.1.0
django-colorfield==0.3.2
django-cors-headers==3.6.0
django-filter==2.4.0
django-url-filter==0.3.15
django-guardian==2.3.0
django-location-field==2.1.0
django-minio-storage==0.3.10
django-oauth-toolkit==1.3.3
django-redis==4.12.1
django-timezone-field==4.0
django-treenode==0.15.0
django-dbbackup==3.3.0
django-storages==1.10.1
django-auditlog==1.0.0
djangorestframework==3.12.2
djangorestframework-datatables==0.5.2
djangorestframework-simplejwt==4.6.0
djangorestframework-word-filter==0.2.0
django-soft-delete==0.9.21
drf-yasg==1.20.0
openapi-codec==1.3.2

# ***********************************
celery==4.4.7   # To update to 5.0
redis==3.5.3
paramiko==2.7.2
Pillow==7.0.0
#psycopg2==2.8.6
psycopg2-binary==2.8.6

python-jose==3.2.0
# python-keycloak==0.19.0
# python-keycloak-client==0.2.2
# django-keycloak==0.1.2.dev0
# dry-rest-permissions==0.1.10  # To removed

sentry-sdk==0.19.1
# websocket-client==0.57.0
xlrd==1.2.0
openpyxl==3.0.5
django-excel==0.0.10
uuid==1.30

gunicorn==20.1.0
boto3==1.16.6
gevent==21.1.2
psycogreen==1.0.2

channels
channels_redis
daphne
channels[daphne]
Twisted[tls,http2]


WeasyPrint==52.5
xlwt==1.3.0
django-admin-autocomplete-filter==0.6.1
requests
numpy==1.21.6
